remove misc
